requirements.txt
pandas==1.3.4
Pillow==9.3.0
streamlit==1.8.1
